+++ flitter/render/helpers.pyx
@@ -9,7 +9,7 @@ from libc.math cimport acos, sqrt
 DEF TwoPI = 6.283185307179586
 
 
-cdef double turn_angle(float x0, float y0, float x1, float y1, float x2, float y2):
+cdef double turn_angle(double x0, double y0, double x1, double y1, double x2, double y2):
     cdef double xa=x1-x0, ya=y1-y0, xb=x2-x1, yb=y2-y1
     cdef double la=sqrt(xa*xa + ya*ya), lb=sqrt(xb*xb + yb*yb)
     if la == 0 or lb == 0:
